more work on buildbar
